pyproject.toml
[project]
name = "manimizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "langchain>=0.3.25",
    "langchain-azure-ai>=0.1.3",
    "langgraph>=0.4.8",
    "manim>=0.19.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
]
